requirements.txt
wandb
numpy
pytorch==2.1.0
torchdiffeq
torchmetrics
torch-fidelity
albumentations
pytorch_lightning
matplotlib
omegaconf
Pillow
einops
open_clip_torch
lpips
argparse
torchvision
pytorch-fid
transformers
webdataset
xformers
